+++ packages/sample-app/src/@types/dynamic-plugin-sdk.d.ts
@@ -1,10 +1,6 @@
-import '@openshift/dynamic-plugin-sdk';
+import type { SupportedCustomProperties } from '../types';
 
 declare module '@openshift/dynamic-plugin-sdk' {
-  interface PluginCustomProperties {
-    sampleApp: {
-      /** What kind of lunch the plugin author wants to eat */
-      lunch: string;
-    };
-  }
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+  interface PluginCustomProperties extends SupportedCustomProperties {}
 }