+++ packages/sample-plugin/src/test-component.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
 
-const TestComponent: React.FC = () => <span>Greetings from sample plugin</span>;
+const TestComponent: React.FC = () => <span>Component from sample plugin</span>;
 
 export default TestComponent;